TodoDetail.jsをコンポーネントに分割
